requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
anyconfig==0.13.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
ax-platform==0.2.5.1
Babel==2.12.1
botorch==0.6.4
build==0.10.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer==2.1.1
click==8.0.4
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
distro==1.8.0
docutils==0.20.1
fabio==2023.6.0
filelock==3.9.0
fire==0.5.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
fvcore==0.1.5.post20221221
gevent==23.7.0
gevent-websocket @ file:///tmp/build/80754af9/gevent-websocket_1614094053512/work
google-auth==2.16.0
google-auth-oauthlib==0.4.6
gpytorch==1.9.1
greenlet==2.0.2
grpcio==1.43.0
h5py==3.9.0
hdf5plugin==4.1.3
huggingface-hub==0.12.0
hydra-core==1.2.0
idna==3.4
imageio==2.25.0
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==6.0.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linear-operator==0.3.0
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
msgpack==1.0.4
multidict==6.0.4
multipledispatch==0.6.0
munch==4.0.0
networkx==3.0
nltk==3.7
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.535.77
nvitop==1.2.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parseq==2023.5.0
Pillow==9.2.0
pip-tools==7.0.0
platformdirs==3.0.0
plotly==5.13.0
portalocker==2.7.0
protobuf==3.20.1
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycodestyle==2.10.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.2
Pygments==2.15.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
ray==1.13.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
shapely==2.0.1
silx==1.1.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.0.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.9.0
tenacity==8.2.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.6.12
tomli==2.0.1
torch==1.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.19.0
Werkzeug==2.2.2
yacs==0.1.8
yarl==1.8.2
zipp==3.13.0
zope.event==5.0
zope.interface==6.0
